update sections in admin interface

+++ views/resumeTable.php
@@ -5,7 +5,7 @@
             <h2>Your <?php echo $sectionTitle; ?></h2>
         </div>
     </div>
-    
+    <br>
     <form method="post">
         <table>
             <tr class="headTable">
@@ -23,28 +23,39 @@
             </tr>
 
             <?php
-            $resumeId = $resume->getId();
-            echo "<tr class=\"contentTable\">";
-            echo "<td>"
-                    . "<a href=\"admin.php?action=showLine&table=resume&id=$resumeId\">"
-                        . "<div class=\"button\">Edit</div>"
-                    . "</a>"
-                    . "</td>"
-                    . "<td>"
-                    . "<a href=\"admin.php?action=deleteDefaultLine\">"
-                        . "<div class=\"button\">Delete</div>"
-                    . "</a>"
-                . "</td>";
-            echo $resume->toString();
-            echo "</tr>";
+            if(isset($resume)) {
+                $resumeId = $resume->getId();
+                echo "<tr class=\"contentTable\">";
+                echo "<td>"
+                        . "<a href=\"admin.php?action=editResume&id=$resumeId\">"
+                            . "<div class=\"button\">Edit</div>"
+                        . "</a>"
+                        . "</td>"
+                        . "<td>"
+                        . "<a href=\"admin.php?action=deleteDefaultLine\">"
+                            . "<div class=\"button\">Delete</div>"
+                        . "</a>"
+                    . "</td>";
+                echo $resume->toString();
+                echo "</tr>";
+            }
             ?>
         </table>
         <?php
             echo "<div class = \"link\">"
-                    . "<a id = \"InsertData\" href = \"admin.php?action=insertInBase\">"
+                    . "<a id = \"InsertData\" href = \"admin.php?action=insertResume&id=$sectionId\">"
                         . "<div class = \"link-div\">Insert</div>"
                     . "</a>"
                 . "</div>";
         ?>
     </form>
+    <?php
+    if (isset($dViewError) && count($dViewError)>0) {
+        foreach($dViewError as $error) {
+        ?>
+        <div class="alert">- <?php echo $error; ?></div>
+        <?php
+        }
+    }
+    ?>
 </div>
